update calendrier & plat components
