+++ projet-cantine/src/app/calendrier/calendrier.component.html
@@ -1,9 +1,11 @@
-<ul class="list-group">
+<!--<ul class="list-group">
     <li class="text-center list-group-item li-title"><strong>{{ dayName }}</strong></li>
-     <app-plat dishName="Burger"></app-plat>
-     <app-plat dishName="Pate"></app-plat>
-     <app-plat dishName="Tarte"></app-plat>
-     <app-plat dishName="Salade"></app-plat>
-     <app-plat dishName="Tomate"></app-plat>
-     <app-plat dishName="Pain au chocolat"></app-plat>
-  </ul>
+    <app-plat dishName="Burger"></app-plat>
+    <app-plat dishName="Pate"></app-plat>
+    <app-plat dishName="Tarte"></app-plat>
+    <app-plat dishName="Salade"></app-plat>
+    <app-plat dishName="Tomate"></app-plat>
+    <app-plat dishName="Pain au chocolat"></app-plat>
+</ul>-->
+
+
