+++ projet-cantine/src/app/infos-modal/infos-modal.component.html
@@ -26,7 +26,7 @@
       A la suite d'une commande, vous recevrez un e-mail de confirmation.
       <br>
       <span class="text-danger">Attention!</span> Ne prenez pas cela pour acquis, la cantinière peut annuler votre commande en cas de problème.
-      Mais ne vous inquiétez pas, vosu serez informé par e-mail.
+      Mais ne vous inquiétez pas, vous serez informé par e-mail.
     </p>
   </div>
 </div>
