+++ projet-cantine/src/app/menu/menu.component.html
@@ -18,9 +18,6 @@
         <li>
             <a href="#" (click)="openFormModalHistorique()"><i class="fas fa-book"></i><span> Historique</span></a>
         </li>
-        <li>
-            <a href="#" (click)="openFormModalInfos()"><i class="fas fa-info"></i><span> Informations pratiques</span></a>
-        </li>
     </ul>
     <div class="gestion">
         <h4>Gestion</h4>
