+++ projet-cantine/src/app/plat/plat.component.html
@@ -1,7 +1,8 @@
-<li class="list-group-item d-flex justify-content-between">
-  <span>{{dishName}}</span><div>
+<!--<li class="list-group-item d-flex justify-content-between">
+  <span>{{dishName}}</span>
+    <div>
         <button type="button" class="btn btn-primary btn-sm mr-1" (click)="openFormModalPlat()"><i class="fas fa-info"></i></button>
         <button type="button" class="btn btn-success btn-sm" (click)="addFoodItem()"><i class="fas fa-plus"></i></button>
     </div>
-</li>
+</li>-->
