+++ projet-cantine/src/app/plat/plat.component.ts
@@ -4,14 +4,23 @@ import {NgbModal} from '@ng-bootstrap/ng-bootstrap';
 
 @Component({
   selector: 'app-plat',
-  templateUrl: './plat.component.html',
+  template: `
+      <li *ngFor="let plat of plats" class="list-group-item d-flex justify-content-between">
+        <span>{{ plat }}</span>
+        <div>
+          <button type="button" class="btn btn-primary btn-sm mr-1" (click)="openFormModalPlat()"><i class="fas fa-info"></i></button>
+          <button type="button" class="btn btn-success btn-sm" (click)="addFoodItem()"><i class="fas fa-plus"></i></button>
+        </div>
+      </li>
+      `,
+  /*templateUrl: './plat.component.html',*/
   styleUrls: ['./plat.component.css']
 })
 export class PlatComponent implements OnInit {
 
   @Input() dishName: string;
   constructor(private modalService: NgbModal) { }
-
+  plats = ['Burger', 'Pate', 'Tarte', 'Salade', 'Tomate', 'Pain au chocolat', 'Patate', 'Oignon'];
   ngOnInit() {
   }
     openFormModalPlat() {
